........
7.2: move $DEFINE WITH_LIFECYCLES to Delphi XE 7 because it gets used there
........
bcd getters for ADO
........
7.2* ADO first changes according the Variant issues of FPC -> use the ActiveX tagVARIANT instead resolves all known date issues and allows access to the unsigned gigint fields
........
SQLite a hack of currency ranges using int affinity
........
7.2* move the tagVar define to XE+
........
7.2* fix weired TestInsertNumbers -> if fieldTypes are widened because of missing TField support -> fallback to max values the provider does support (FPC/D7..D2010)
........
7.2* SP test fixes with the defines
........
7.2* first ADO callable fixes for the FPC (the variantmanager can't handle 1,2byte ordinals, date values nor decimals from the OleVariants -> raises Variant-Errors)
the VirtualOpen from OutParams need to be changed too
........
skip sqlite FmtBCD tests..
........
unsigned TiniInts are mapped to ftWord if the compiler supports it
........
7.2* ADO get all test passed with FPC 3.0.4 hope 2.6 will compile too?
........
7.2?: Changes of Whitewind to have Zeos 7.3 compile on Linux. See http://zeoslib.sourceforge.net/viewtopic.php?p=117995#p117995
........
fix PGCASH in ColumnsToJSON
........
7.2 complete the ADO resultset jurney for full old delphi and FPC support
........
ADO change descendant-type of TZAdoResultSet to TZAbstractReadOnlyResultSet for faster access (no virtual functions used)
........
get d2009 compiled again
........


git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/trunk@5696 d681b350-854c-439c-a81d-c0ce20710203

+++ src/dbc/ZDbcASAUtils.pas
@@ -1005,11 +1005,11 @@ begin
         begin
           case Self.GetFieldSqlType(Index) of
             stAsciiStream:
-              SetFieldType(TempSQLDA, 0, DT_LONGVARCHAR, Min( BlockSize, Length));
+              SetFieldType(TempSQLDA, 0, DT_LONGVARCHAR, Min( Int64(BlockSize), Int64(Length)));
             stUnicodeStream:
-              SetFieldType(TempSQLDA, 0, DT_LONGNVARCHAR, Min( BlockSize, Length));
+              SetFieldType(TempSQLDA, 0, DT_LONGNVARCHAR, Min( Int64(BlockSize), Int64(Length)));
             stBinaryStream:
-              SetFieldType(TempSQLDA, 0, DT_LONGBINARY, Min( BlockSize, Length));
+              SetFieldType(TempSQLDA, 0, DT_LONGBINARY, Min( Int64(BlockSize), Int64(Length)));
             else
               sqlType := DT_FIXCHAR;
           end;
@@ -1032,7 +1032,7 @@ begin
             if ( sqlind^ = 0 ) or ( RD = Length) then
               break;
             Inc( Offs, PZASABlobStruct( sqlData)^.stored_len);
-            sqllen := Min( BlockSize, Length-Rd);
+            sqllen := Min( Int64(BlockSize), Int64(Length-Rd));
           end;
           if Rd <> Length then
             CreateException( 'Could''nt complete BLOB-Read');
